MFB-1676: put shared test scaffolding under programs/programs/testing/
The helpers carved out of the split modules were sitting in programs/framework/tests/
even though every consumer is a program test. They move to programs/programs/testing/,
which holds fixtures and base classes and no test cases:

  pe_integration.py           PE household fixtures + PeIntegrationTestCase
  pe_input_test_base.py       payload assertions, used by 10 families' TX tests
  current_benefit_fixtures.py a screen that receives a named benefit

Named 'testing' rather than 'test_helpers' because the registry walks
programs.programs and skips anything prefixed 'test_', which would have made these
unimportable from the tests that need them.

Two files had near-identical names that said nothing about their contents:
integration_test_helpers.py is now testing/pe_integration.py, and
test_integration_helpers.py is now framework/tests/test_pe_integration_harness.py.

Also:
- registered_subclasses / state_codes / distinct_state_codes lose their leading
  underscores; they are a module's public API, imported from three others. The
  module moves to cross_white_label/test_helpers.py, its only consumers being
  families there.
- The two IL mixins move to white_labels/il/, each with the four and two consumer
  classes that use them. cross_white_label/csfp/il.py had imported one and never
  used it.
- IL's family_care, moms_and_babies and aca_adults join cross_white_label/medicaid/,
  where il_all_kids already was.
- TX's four Medicaid tests move from medicaid/tests/test_tx.py into the four
  subgroups holding those programs; TX has no plain Medicaid program.
- test_presumptive_eligibility_resolution.py keeps the three tests of the
  resolution seam; SUN Bucks' exclusions and NFP's WIC bypass move to those
  programs.
- programs/framework/helpers.py stays: screener/views.py reads
  STATE_MEDICAID_OPTIONS for calculation ordering, so it is used outside the
  program tree.

Registry keys and class names identical to main (121 PE / 118 MFB); 2690 test
methods present by name; 3331 tests pass.
