+++ programs/framework/tests/test_pe_integration_harness.py
@@ -15,7 +15,7 @@ change, not about the harness. Re-record after reviewing the new body.
 
 import pytest
 
-from .integration_test_helpers import (
+from programs.programs.testing.pe_integration import (
     PeIntegrationTestCase,
     add_income,
     add_member,
