+++ programs/programs/cross_white_label/aca/tests/test_tx.py
@@ -3,7 +3,7 @@
 from programs.framework.pe_dependencies.constants import MAIN_TAX_UNIT
 from programs.programs.cross_white_label.aca.tx import TxAca
 from integrations.clients.policyengine.policy_engine import pe_input
-from programs.framework.tests.pe_input_test_base import TxPeInputTestBase
+from programs.programs.testing.pe_input_test_base import TxPeInputTestBase
 from django.test import TestCase
 from programs.programs.cross_white_label.aca.base import Aca
 from programs.framework.pe_dependencies.household import TxStateCodeDependency
