+++ programs/programs/cross_white_label/csfp/il.py
@@ -1,5 +1,4 @@
 from programs.framework.base import ProgramCalculator, Eligibility, MemberEligibility
-from programs.framework.mixins import IlTransportationMixin
 
 
 class IlCommoditySupplementalFoodProgram(ProgramCalculator):
