+++ programs/programs/cross_white_label/eitc/tests/test_tx.py
@@ -6,7 +6,7 @@ from screener.models import IncomeStream
 from programs.framework.pe_dependencies.constants import MAIN_TAX_UNIT
 from screener.models import Screen
 from integrations.clients.policyengine.policy_engine import pe_input
-from programs.framework.tests.pe_input_test_base import TxPeInputTestBase
+from programs.programs.testing.pe_input_test_base import TxPeInputTestBase
 from django.test import TestCase
 from programs.programs.cross_white_label.eitc.tx import TxEitc
 from programs.framework.pe_dependencies import household
