+++ programs/programs/cross_white_label/lifeline/tests/test_tx.py
@@ -2,7 +2,7 @@
 
 from programs.programs.cross_white_label.lifeline.tx import TxLifeline
 from integrations.clients.policyengine.policy_engine import pe_input
-from programs.framework.tests.pe_input_test_base import TxPeInputTestBase
+from programs.programs.testing.pe_input_test_base import TxPeInputTestBase
 from django.test import TestCase
 from programs.programs.cross_white_label.lifeline.base import Lifeline
 from programs.framework.pe_dependencies.household import TxStateCodeDependency
