+++ programs/programs/cross_white_label/medicaid/chip/il.py
@@ -1,7 +1,7 @@
 from programs.framework.base import MemberEligibility, ProgramCalculator, Eligibility
 from programs.framework.helpers import medicaid_eligible
 import programs.framework.eligibility_messages as messages
-from programs.framework.mixins import IlMedicaidFplIncomeCheckMixin
+from programs.programs.white_labels.il.medicaid_fpl_mixin import IlMedicaidFplIncomeCheckMixin
 
 
 class AllKids(ProgramCalculator, IlMedicaidFplIncomeCheckMixin):
