+++ programs/programs/cross_white_label/medicaid/chip/tests/test_tx.py
@@ -2,7 +2,7 @@
 
 from programs.programs.cross_white_label.medicaid.chip.tx import TxChip
 from integrations.clients.policyengine.policy_engine import pe_input
-from programs.framework.tests.pe_input_test_base import TxPeInputTestBase
+from programs.programs.testing.pe_input_test_base import TxPeInputTestBase
 from django.test import TestCase
 from unittest.mock import MagicMock
 from programs.programs.cross_white_label.medicaid.base import Medicaid
