+++ programs/programs/cross_white_label/ssi/tests/test_tx.py
@@ -2,7 +2,7 @@
 
 from programs.programs.cross_white_label.ssi.tx import TxSsi
 from integrations.clients.policyengine.policy_engine import pe_input
-from programs.framework.tests.pe_input_test_base import TxPeInputTestBase
+from programs.programs.testing.pe_input_test_base import TxPeInputTestBase
 from django.test import TestCase
 from programs.programs.cross_white_label.ssi.base import Ssi
 from programs.framework.pe_dependencies.household import TxStateCodeDependency
