+++ programs/programs/cross_white_label/wic/tests/test_tx.py
@@ -3,7 +3,7 @@
 from screener.models import HouseholdMember
 from programs.programs.cross_white_label.wic.tx import TxWic
 from integrations.clients.policyengine.policy_engine import pe_input
-from programs.framework.tests.pe_input_test_base import TxPeInputTestBase
+from programs.programs.testing.pe_input_test_base import TxPeInputTestBase
 from django.test import TestCase
 from programs.framework.pe_dependencies.household import TxStateCodeDependency
 from programs.programs.cross_white_label.wic.base import Wic
